add removepropery module

+++ Airbnb Clone/Server/controllers/propertycontroller.js
@@ -69,6 +69,11 @@ module.exports.updateproperty = async (req, res) => {
     try {
         if(req.file){
             const property = await Property.findById(req.params.id);
+
+            if(property.ownerId.toString() !== req.user._id){
+                return res.status(400).json({ msg: 'Not authorized to update this property!!', status: 0, response: 'error' });
+            }
+
             var imgpath = path.join(__dirname,'..',property.image);
             if(fs.existsSync(imgpath)){
                 fs.unlinkSync(imgpath); //delete image
@@ -85,6 +90,11 @@ module.exports.updateproperty = async (req, res) => {
         }
         else{
             const property = await Property.findById(req.params.id);
+
+            if(property.ownerId.toString() !== req.user._id){
+                return res.status(400).json({ msg: 'Not authorized to update this property!!', status: 0, response: 'error' });
+            }
+
             req.body.image = property.image;
             req.body.updateAt = moment().format('LLL');
             const updateproperty = await Property.findByIdAndUpdate(req.params.id, req.body, {new: true});
@@ -106,8 +116,19 @@ module.exports.removeproperty = async (req, res) => {
     try {
         const property = await Property.findById(req.params.id);
         var imgpath = path.join(__dirname, '..', property.image);
-        fs.unlinkSync(imgpath);
-        const rmvproperty = await Property.findByIdAndDelete(req.params.id,{ownerId: req.user._id});
+        if(property){
+            if(property.ownerId.toString() !== req.user._id){
+                return res.status(400).json({ msg: 'Not authorized to delete this property!!', status: 0, response: 'error' });
+            }
+            else{
+                fs.unlinkSync(imgpath);
+            }
+        }
+        else{
+            return res.status(400).json({ msg: 'Property is not found!!', status: 0, response: 'error' });
+        }
+
+        const rmvproperty = await Property.findByIdAndDelete(req.params.id);
         if(rmvproperty){
             return res.status(200).json({ msg: 'Property is deleted successfully', status: 1, response: 'success' });
         }
@@ -120,3 +141,4 @@ module.exports.removeproperty = async (req, res) => {
         return res.status(400).json({ msg: 'Somthing went wrong', status: 0, response: 'error' });
     }
 }
+
